Move job handling to a separate module
